requirements.txt
langchain-core>=0.1.0
langchain-huggingface>=0.0.7
huggingface_hub>=0.20.0

pytest>=7.4.0
python-dotenv>=1.0.0
